drivers: flash: Add driver write-block-size requirement to Kconfig
This path introduce fixed write-block-size requirement of flash driver
as the Kconfig constant integer.
This value can defer from hardware alignment requirement
(as for nrf5x is). It is not expected that this value changes in
run-time so macro value was chosen because it allows to use compile time
optimization.

As the driver has certain requirement for alignment
on the writing, it is necessary to export this value for upper modules
which needs to know the write-block-size (for instance NFFS needs this).

Appropriate values was specified for nRF5x, STM32 and QMSI SoCs.

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>