drivers: flash: add flash_simulator driver
This commit adds a flash driver implementation that writes to RAM and
exports statistics through stats.h. It can be used to simulate flash
memory for testing purposes. When the simulator is enabled a common
dts overlay file is pulled in automatically and a default device name
is exported via dts fixups for Nordic SoCs and qemu_cortex_m3 and
qemu_x86 boards.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>
Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>