arch: arm: cortex-m: define vector relay tables for bootloader only
If an image is not a bootloader image (such as an MCUboot image)
but it is a standard Zephyr firmware, chain-loadable by a
bootloader, then this image will not need to relay IRQs itself.
In this case SW_VECTOR_RELAY should restrict to setting the
vector table pointer in RAM so the parent image can forward the
interrupts to  it.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>